requirements.txt
asyncio
aiosqlite
sqlalchemy
strawberry-graphql
fastapi
uvicorn